test(architecture): extract fastify app route stubs
